# Only run simplecov in CI env