--- a/spec/spec_helper.rb
+++ b/spec/spec_helper.rb
@@ -1,4 +1,4 @@
-if ENV['CI']
+if ENV['COVERAGE']
   require 'simplecov'
   SimpleCov.start 'rails' do
     add_filter '/config/'