**Why**:
* When running `rpsec spec/ --profile`, i18n-tasks was slowest file
* Ran 3 times, each time file took 35+ seconds to run
* Upon investigation, this appears to relate to Simplecov
* See https://github.com/glebm/i18n-tasks/issues/221
* Also https://github.com/colszowka/simplecov/issues/404
* Since there is no proposed solution at the moment (slowness appears to
  be coming from `Coverage`, which is part of the Ruby stdlib), the
  short term solution is to only run Simplecov on Travis, where required
* New timing for i18n-tasks: under 3 seconds!!!

Before:

```
Top 10 slowest example groups:
  I18n
    17.89 seconds average (35.78 seconds / 2 examples) ./spec/i18n_spec.rb:4
```
After:

```
Top 10 slowest example groups:
  I18n
    1.35 seconds average (2.71 seconds / 2 examples) ./spec/i18n_spec.rb:4
```