@jessieay 2017-02-16
@zachmargolis I think you make good points. Updated name at 3a7c61c